+++ python/cudf/tests/test_rolling.py
@@ -45,10 +45,17 @@ def test_rollling_series_basic(data, agg, nulls, center):
     for window_size in range(1, len(data)+1):
         for min_periods in range(1, window_size+1):
             assert_eq(
-                getattr(psr.rolling(window_size, min_periods, center), agg)().fillna(-1),
-                getattr(gsr.rolling(window_size, min_periods, center), agg)().fillna(-1),
+                getattr(
+                    psr.rolling(window_size, min_periods, center),
+                    agg
+                )().fillna(-1),
+                getattr(
+                    gsr.rolling(window_size, min_periods, center),
+                    agg
+                )().fillna(-1),
                 check_dtype=False
             )
+
 
 @pytest.mark.parametrize(
     'data',
@@ -56,7 +63,10 @@ def test_rollling_series_basic(data, agg, nulls, center):
         {'a': [], 'b': []},
         {'a': [1, 2, 3, 4], 'b': [1, 2, 3, 4]},
         {'a': [1, 2, 4, 9, 9, 4], 'b': [1, 2, 4, 9, 9, 4]},
-        {'a': np.array([1, 2, 4, 9, 9, 4]), 'b': np.array([1.5, 2.2, 2.2, 8.0, 9.1, 4.2])}
+        {
+            'a': np.array([1, 2, 4, 9, 9, 4]),
+            'b': np.array([1.5, 2.2, 2.2, 8.0, 9.1, 4.2])
+        }
     ]
 )
 @pytest.mark.parametrize(
@@ -90,7 +100,13 @@ def test_rolling_dataframe_basic(data, agg, nulls, center):
     for window_size in range(1, len(data)+1):
         for min_periods in range(1, window_size+1):
             assert_eq(
-                getattr(pdf.rolling(window_size, min_periods, center), agg)().fillna(-1),
-                getattr(gdf.rolling(window_size, min_periods, center), agg)().fillna(-1),
+                getattr(
+                    pdf.rolling(window_size, min_periods, center),
+                    agg
+                )().fillna(-1),
+                getattr(
+                    gdf.rolling(window_size, min_periods, center),
+                    agg
+                )().fillna(-1),
                 check_dtype=False
             )